Marge 'MoveDelete' feature

+++ pyPromptChecker/lib/io.py
@@ -1,3 +1,5 @@
+# -*- coding: utf-8 -*-
+
 import csv
 import json
 import os
@@ -8,8 +10,12 @@ from PyQt6.QtWidgets import QApplication
 
 
 def export_json(target_json, filepath):
-    with open(filepath, 'w') as f:
-        json.dump(target_json, f, sort_keys=True, indent=4, ensure_ascii=False)
+    try:
+        with open(filepath, 'w') as f:
+            json.dump(target_json, f, sort_keys=True, indent=4, ensure_ascii=False)
+            return True, None
+    except Exception as e:
+        return 'Error occurred during writing JSON.', e
 
 
 def import_json():
@@ -27,20 +33,24 @@ def import_model_list(filepath):
 
 
 def image_copy_to(source, destination, is_move=False):
-    try:
-        shutil.copy(source, destination)
-        if os.path.exists(destination) and is_move:
-            os.remove(source)
-            return True
-        return False
-    except Exception as e:
-        return False, e
+    destination_path = os.path.join(destination, os.path.basename(source))
+    if not os.path.exists(destination_path):
+        try:
+            shutil.copy(source, destination)
+            if os.path.exists(destination) and is_move:
+                os.remove(source)
+                return True, None
+            return True, None
+        except Exception as e:
+            return 'Error occurred moving/copying files.', e
+    else:
+        return 'The same files already exists in destination.', 'AlreadyExistsError'
 
 
-def clear_trash_bin(dirpath):
-    file_list = os.listdir(dirpath)
+def clear_trash_bin(directory_path):
+    file_list = os.listdir(directory_path)
     for filename in file_list:
-        filepath = os.path.join(dirpath, filename)
+        filepath = os.path.join(directory_path, filename)
         try:
             os.remove(filepath)
         except Exception as e:
@@ -48,8 +58,8 @@ def clear_trash_bin(dirpath):
     return True
 
 
-def is_directory_empty(dirpath):
-    file_list = os.listdir(dirpath)
+def is_directory_empty(directory_path):
+    file_list = os.listdir(directory_path)
     return len(file_list) == 0
